moved create_new_rewrite to dialogs.py
